refactor(api-gateway): Phase 3 relocate proxy plumbing to Infrastructure/Proxy/
- Move YARP/Ocelot/SwaggerAggregation modules + GatewayProvider switch into
  Infrastructure/Proxy/ with matching namespaces; delete the empty Gateway/.
- Update Program.cs + tests (GatewayProviderExtensionsTests, GatewaySpecTransformerTests,
  GatewayWebApplicationFactory) to the new namespaces; behavior unchanged.
